+++ hardpy/common/config.py
@@ -79,6 +79,7 @@ class FrontendConfig(BaseModel):
     manual_collect: bool = False
     measurement_display: bool = True
     test_history: bool = False
+    auto_scroll: bool = False
     modal_result: ModalResultConfig = Field(default_factory=lambda: ModalResultConfig())
     reports_storage_menu: ReportsStorageMenuConfig = Field(
         default_factory=lambda: ReportsStorageMenuConfig(),